// src/main/types.ts
export interface IThriftProject {
    type: 'ThriftProject'

    // Root directory for the project
    rootDir: string

    // Output directory (relative to root) for generated code
    outDir: string

    // Source directory (relative to root) for thrift files
    sourceDir: string

    // Namespace declared in this project
    namespaces: INamespaceMap

    // Options for rendering this project
    options: IMakeOptions
}

export interface INamespaceFiles {
    [name: string]: Array<IParsedFile>
}

// Map of resolved identifier name to the namespace it represents
export interface INamespacePathMap {
    [resolvedName: string]: INamespacePath
}

export interface INamespacePath {
    type: 'NamespacePath'

    // Scope is the language this namespace belongs to 'js', 'java'...
    scope: string

    // The name of the namespace com.company.package
    name: string

    // The name translated to its result path com/company/package
    path: string
}

// Namespace path to namespace
export interface INamespaceMap {
    [path: string]: INamespace
}

export interface INamespace {
    type: 'Namespace'

    namespace: INamespacePath

    // Files declared as part of this namespace
    files: ResolvedFileMap

    // Identifiers defined in this namespace
    exports: IFileExports

    // Map of namespaces used by this file
    includedNamespaces: INamespacePathMap

    // Data/services defined in this namespace
    constants: Array<ConstDefinition | EnumDefinition>
    typedefs: Array<TypedefDefinition>
    structs: Array<StructDefinition>
    unions: Array<UnionDefinition>
    exceptions: Array<ExceptionDefinition>
    services: Array<ServiceDefinition>
}

export interface IRenderState {
    // Options for this render
    options: IMakeOptions

    // The current namespace being processed
    currentNamespace: INamespace

    // Current statements being processed
    currentDefinitions: IFileExports

    // The current Thrift project
    project: IThriftProject
}

export type CompileTarget = 'apache' | 'thrift-server'

export interface IMakeOptions {
    // Root to resolve outDir and sourceDir from
    rootDir: string

    // Where to put generated TypeScript
    outDir: string

    // Where to find source thrift
    sourceDir: string

    // Files to generate from
    files: Array<string>

    // What core libs are you compiling for?
    target: CompileTarget

    // What namespace do we fallback to? Set to 'none' to not fallback.
    fallbackNamespace: string

    // What is the library to import thrift from
    // Defaults to 'thrift' for target = 'apache'
    // Defaults to '@creditkarma/thrift-server-core' for target = 'thrift-server'
    library: string

    // Should we render strict unions?
    strictUnions: boolean

    // Should render strict unions type with underscore
    strictUnionsComplexNames: boolean
}
// ...
export type ResolvedFileMap = IProcessedFileMap<IResolvedFile>
// ...
export type DefinitionType =
    | ConstDefinition
    | StructDefinition
    | UnionDefinition
    | ExceptionDefinition
    | EnumDefinition
    | TypedefDefinition
    | ServiceDefinition

// A map of identifier name to the Thrift definition of that identifier
export interface IFileExports {
    [name: string]: DefinitionType
}
// ...
export interface IResolvedIdentifier {
    name: string
    baseName: string
    pathName?: string
    fullName: string
}